+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ master
 ------
 
 - (`#67 <https://github.com/openscm/scmdata/pull/67>`_) Update the documentation for :meth`ScmRun.append`
+- (`#66 <https://github.com/openscm/scmdata/pull/66>`_) Raise ValueError if index/columns arguments are not provided when instantiating a :class`ScmRun` object with a numpy array. Add ``lowercase_cols`` argument to coerce the column names in CSV files to lowercase
 
 v0.5.0
 ------